# movie-sphere-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegistrationRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    },
    "/auth/authenticate": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthenticationRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/movies": {
      "get": {
        "tags": ["Movie/Admin"],
        "operationId": "searchMovies",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PageResponseMovie" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Movie/Admin"],
        "operationId": "addMovie",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "array", "items": { "type": "string" } }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Movie" }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Movie/Admin"],
        "operationId": "deleteMultipleMovies",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": { "type": "integer", "format": "int64" }
              }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/admin/movies/{imdbId}": {
      "get": {
        "tags": ["Movie/Admin"],
        "operationId": "getMovie",
        "parameters": [
          {
            "name": "imdbId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Movie" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Movie/Admin"],
        "operationId": "addMovie_1",
        "parameters": [
          {
            "name": "imdbId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Movie" }
              }
            }
          }
        }
      }
    },
    "/user/movies": {
      "get": {
        "tags": ["Movie/User"],
        "operationId": "getAllMovies",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 5 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PageResponseMovie" }
              }
            }
          }
        }
      }
    },
    "/user/movies/{id}": {
      "get": {
        "tags": ["Movie/User"],
        "operationId": "getMovie_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Movie" }
              }
            }
          }
        }
      }
    },
    "/user/movies/search": {
      "get": {
        "tags": ["Movie/User"],
        "operationId": "searchMovies_1",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 5 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PageResponseMovie" }
              }
            }
          }
        }
      }
    },
    "/admin/movies/{movieId}": {
      "delete": {
        "tags": ["Movie/Admin"],
        "operationId": "deleteMovie",
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    }
  },
  "components": {
    "schemas": {
      "RegistrationRequest": {
        "required": ["email", "firstName", "lastName", "password"],
        "type": "object",
        "properties": {
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string" },
          "password": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "AuthenticationRequest": {
        "required": ["email", "password"],
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": { "token": { "type": "string" } }
      },
      "Movie": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "Title": { "type": "string" },
          "Year": { "type": "string" },
          "Rated": { "type": "string" },
          "Released": { "type": "string" },
          "Runtime": { "type": "string" },
          "Genre": { "type": "string" },
          "Director": { "type": "string" },
          "Writer": { "type": "string" },
          "Actors": { "type": "string" },
          "Plot": { "type": "string" },
          "Language": { "type": "string" },
          "Country": { "type": "string" },
          "Awards": { "type": "string" },
          "Poster": { "type": "string" },
          "Metascore": { "type": "string" },
          "imdbRating": { "type": "string" },
          "imdbVotes": { "type": "string" },
          "imdbID": { "type": "string" },
          "Type": { "type": "string" },
          "DVD": { "type": "string" },
          "BoxOffice": { "type": "string" },
          "Production": { "type": "string" },
          "Website": { "type": "string" },
          "Response": { "type": "string" },
          "Ratings": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Rating" }
          }
        }
      },
      "Rating": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "Source": { "type": "string" },
          "Value": { "type": "string" }
        }
      },
      "PageResponseMovie": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Movie" }
          },
          "page": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" }
        }
      }
    }
  }
}